+++ sdk/hybridcompute/azure-resourcemanager-hybridcompute/src/main/java/com/azure/resourcemanager/hybridcompute/implementation/HybridComputeManagementClientImpl.java
@@ -23,11 +23,16 @@ import com.azure.core.util.polling.LongRunningOperationStatus;
 import com.azure.core.util.polling.PollerFlux;
 import com.azure.core.util.serializer.SerializerAdapter;
 import com.azure.core.util.serializer.SerializerEncoding;
+import com.azure.resourcemanager.hybridcompute.fluent.ExtensionMetadataV2sClient;
 import com.azure.resourcemanager.hybridcompute.fluent.ExtensionMetadatasClient;
+import com.azure.resourcemanager.hybridcompute.fluent.ExtensionPublishersClient;
+import com.azure.resourcemanager.hybridcompute.fluent.ExtensionTypesClient;
+import com.azure.resourcemanager.hybridcompute.fluent.GatewaysClient;
 import com.azure.resourcemanager.hybridcompute.fluent.HybridComputeManagementClient;
 import com.azure.resourcemanager.hybridcompute.fluent.LicenseProfilesClient;
 import com.azure.resourcemanager.hybridcompute.fluent.LicensesClient;
 import com.azure.resourcemanager.hybridcompute.fluent.MachineExtensionsClient;
+import com.azure.resourcemanager.hybridcompute.fluent.MachineRunCommandsClient;
 import com.azure.resourcemanager.hybridcompute.fluent.MachinesClient;
 import com.azure.resourcemanager.hybridcompute.fluent.NetworkProfilesClient;
 import com.azure.resourcemanager.hybridcompute.fluent.NetworkSecurityPerimeterConfigurationsClient;
@@ -36,6 +41,7 @@ import com.azure.resourcemanager.hybridcompute.fluent.PrivateEndpointConnections
 import com.azure.resourcemanager.hybridcompute.fluent.PrivateLinkResourcesClient;
 import com.azure.resourcemanager.hybridcompute.fluent.PrivateLinkScopesClient;
 import com.azure.resourcemanager.hybridcompute.fluent.ResourceProvidersClient;
+import com.azure.resourcemanager.hybridcompute.fluent.SettingsOperationsClient;
 import java.io.IOException;
 import java.lang.reflect.Type;
 import java.nio.ByteBuffer;
@@ -219,6 +225,48 @@ public final class HybridComputeManagementClientImpl implements HybridComputeMan
     }
 
     /**
+     * The ExtensionMetadataV2sClient object to access its operations.
+     */
+    private final ExtensionMetadataV2sClient extensionMetadataV2s;
+
+    /**
+     * Gets the ExtensionMetadataV2sClient object to access its operations.
+     * 
+     * @return the ExtensionMetadataV2sClient object.
+     */
+    public ExtensionMetadataV2sClient getExtensionMetadataV2s() {
+        return this.extensionMetadataV2s;
+    }
+
+    /**
+     * The ExtensionTypesClient object to access its operations.
+     */
+    private final ExtensionTypesClient extensionTypes;
+
+    /**
+     * Gets the ExtensionTypesClient object to access its operations.
+     * 
+     * @return the ExtensionTypesClient object.
+     */
+    public ExtensionTypesClient getExtensionTypes() {
+        return this.extensionTypes;
+    }
+
+    /**
+     * The ExtensionPublishersClient object to access its operations.
+     */
+    private final ExtensionPublishersClient extensionPublishers;
+
+    /**
+     * Gets the ExtensionPublishersClient object to access its operations.
+     * 
+     * @return the ExtensionPublishersClient object.
+     */
+    public ExtensionPublishersClient getExtensionPublishers() {
+        return this.extensionPublishers;
+    }
+
+    /**
      * The OperationsClient object to access its operations.
      */
     private final OperationsClient operations;
@@ -244,6 +292,48 @@ public final class HybridComputeManagementClientImpl implements HybridComputeMan
      */
     public NetworkProfilesClient getNetworkProfiles() {
         return this.networkProfiles;
+    }
+
+    /**
+     * The MachineRunCommandsClient object to access its operations.
+     */
+    private final MachineRunCommandsClient machineRunCommands;
+
+    /**
+     * Gets the MachineRunCommandsClient object to access its operations.
+     * 
+     * @return the MachineRunCommandsClient object.
+     */
+    public MachineRunCommandsClient getMachineRunCommands() {
+        return this.machineRunCommands;
+    }
+
+    /**
+     * The GatewaysClient object to access its operations.
+     */
+    private final GatewaysClient gateways;
+
+    /**
+     * Gets the GatewaysClient object to access its operations.
+     * 
+     * @return the GatewaysClient object.
+     */
+    public GatewaysClient getGateways() {
+        return this.gateways;
+    }
+
+    /**
+     * The SettingsOperationsClient object to access its operations.
+     */
+    private final SettingsOperationsClient settingsOperations;
+
+    /**
+     * Gets the SettingsOperationsClient object to access its operations.
+     * 
+     * @return the SettingsOperationsClient object.
+     */
+    public SettingsOperationsClient getSettingsOperations() {
+        return this.settingsOperations;
     }
 
     /**
@@ -319,15 +409,21 @@ public final class HybridComputeManagementClientImpl implements HybridComputeMan
         this.defaultPollInterval = defaultPollInterval;
         this.subscriptionId = subscriptionId;
         this.endpoint = endpoint;
-        this.apiVersion = "2024-07-10";
+        this.apiVersion = "2024-09-10-preview";
         this.licenses = new LicensesClientImpl(this);
         this.machines = new MachinesClientImpl(this);
         this.licenseProfiles = new LicenseProfilesClientImpl(this);
         this.machineExtensions = new MachineExtensionsClientImpl(this);
         this.resourceProviders = new ResourceProvidersClientImpl(this);
         this.extensionMetadatas = new ExtensionMetadatasClientImpl(this);
+        this.extensionMetadataV2s = new ExtensionMetadataV2sClientImpl(this);
+        this.extensionTypes = new ExtensionTypesClientImpl(this);
+        this.extensionPublishers = new ExtensionPublishersClientImpl(this);
         this.operations = new OperationsClientImpl(this);
         this.networkProfiles = new NetworkProfilesClientImpl(this);
+        this.machineRunCommands = new MachineRunCommandsClientImpl(this);
+        this.gateways = new GatewaysClientImpl(this);
+        this.settingsOperations = new SettingsOperationsClientImpl(this);
         this.privateLinkScopes = new PrivateLinkScopesClientImpl(this);
         this.privateLinkResources = new PrivateLinkResourcesClientImpl(this);
         this.privateEndpointConnections = new PrivateEndpointConnectionsClientImpl(this);